% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_tableau_font_style.R
\name{adjust_tableau_font_style}
\alias{adjust_tableau_font_style}
\title{adjust_tableau_font_style}
\usage{
adjust_tableau_font_style(
  import_files,
  font_style = "Tableau Regular",
  save_location = NULL
)
}
\arguments{
\item{import_files}{selected tableau file or folder, to change its font style. If
it is only a file, make sure it is a twb file.}

\item{font_style}{the name of the font style you want to use (in quotation marks "")}

\item{save_location}{The location to which the adjusted tableau file will be
saved (this needs to be a .twb file). If empty, the adjusted tableau file will be overwritten.}
}
\value{
tableau file with the correct font style.
}
\description{
Adjust the font style of a tableau file
}
