% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_tableau_size.R
\name{adjust_tableau_size}
\alias{adjust_tableau_size}
\title{adjust_tableau_size}
\usage{
adjust_tableau_size(
  import_files,
  save_location = NULL,
  maxheight = NULL,
  maxwidth = NULL,
  minheight = NULL,
  minwidth = NULL
)
}
\arguments{
\item{import_files}{selected tableau file or folder, to change its size. If
it is only one file, make sure it is a twb file.}

\item{save_location}{The location to which the adjusted tableau file will be
saved (this needs to be a .twb file). If empty, the adjusted tableau file will be overwritten.}

\item{maxheight}{Max height for tableau file, if NULL use system variables.}

\item{maxwidth}{Max width for tableau file, if NULL use system variables.}

\item{minheight}{Min height for tableau file, if NULL use system variables.}

\item{minwidth}{Min width for tableau file, if NULL use system variables.}
}
\value{
tableau file with the correct size.
}
\description{
Adjust the size of selected tableau file/files.
}
\seealso{
Other xml: 
\code{\link{check_dashboard_names}()},
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
