% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dashboard_names.R
\name{check_dashboard_names}
\alias{check_dashboard_names}
\title{check_dashboard_names}
\usage{
check_dashboard_names(import_files)
}
\arguments{
\item{import_files}{selected tableau file or folder,
to check the names of the dashboards.
If it is only one file, make sure it is a twb file.}
}
\value{
the wrongly named dashboards
}
\description{
Check the names of dashboards in tableau, according to the style guide.
The wrong dashboards will be returned together with the reason why it is a
wrong name
}
\seealso{
Other xml: 
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
