% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_group.R
\name{delete_group}
\alias{delete_group}
\title{Delete Group}
\usage{
delete_group(tableau, group_id, api_version = 3.19)
}
\arguments{
\item{tableau}{A list containing the Tableau authentication variables: `base_url`, `token`, `user_id`, and `site_id`.}

\item{group_id}{The ID of the group to delete.}

\item{api_version}{The API version to use (default: 3.19).}
}
\description{
Deletes the specified group from the site.
}
