% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tableau_crosstab_excel.R
\name{download_tableau_crosstab_excel}
\alias{download_tableau_crosstab_excel}
\title{Download Tableau view crosstab as Excel}
\usage{
download_tableau_crosstab_excel(
  base_url,
  site_id,
  token,
  view_id,
  path_to_save,
  api_version = 3.16
)
}
\arguments{
\item{base_url}{Tableau server URL}

\item{site_id}{The site ID}

\item{token}{API token}

\item{view_id}{The ID of the view to download}

\item{path_to_save}{The directory to write the crosstab Excel file to}

\item{api_version}{The API version; default set to 3.16}
}
\description{
Download Tableau view crosstab as Excel
}
\seealso{
Other tableau rest api: 
\code{\link{authenticate_server}()},
\code{\link{download_filtered_tableau_image}()},
\code{\link{download_workbooks_server_pdf}()},
\code{\link{download_workbooks_server_powerpoint}()},
\code{\link{download_workbooks_server}()},
\code{\link{get_server_connected_apps}()},
\code{\link{get_server_datasources}()},
\code{\link{get_server_groups}()},
\code{\link{get_server_info}()},
\code{\link{get_server_jobs}()},
\code{\link{get_server_projects}()},
\code{\link{get_server_refresh_tasks}()},
\code{\link{get_server_schedules}()},
\code{\link{get_server_users}()},
\code{\link{get_server_views}()},
\code{\link{get_server_workbooks}()}
}
\concept{tableau rest api}
