% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view_filtered.R
\name{escape_special_chars}
\alias{escape_special_chars}
\title{Escape characters for url}
\usage{
escape_special_chars(url)
}
\arguments{
\item{url}{the url}
}
\value{
escaped string
}
\description{
Escape characters for url
}
