% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actions.R
\name{get_actions}
\alias{get_actions}
\title{Get Workbook actions}
\usage{
get_actions(wb)
}
\arguments{
\item{wb}{The path to the Tableau workbook [.twb].}
}
\value{
Dataframe containing the workbook actions.
}
\description{
Get Workbook actions
}
\examples{
\dontrun{
# Get Workbook actions
actions <- get_actions(wb = "path/to/workbook.twb")
head(actions)
}
}
\seealso{
Other xml: 
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
