% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_for_user.R
\name{get_groups_for_user}
\alias{get_groups_for_user}
\title{Get Groups for a User}
\usage{
get_groups_for_user(
  tableau,
  api_version = 3.19,
  user_id,
  page_size = 100,
  page_number = 1,
  include_metadata = FALSE
)
}
\arguments{
\item{tableau}{A list containing the Tableau authentication variables: `base_url`, `token`, `user_id`, and `site_id`.}

\item{api_version}{The API version to use (default: 3.19).}

\item{user_id}{The ID of the user whose group memberships are listed.}

\item{page_size}{(Optional) The number of items to return in one response. The minimum is 1. The maximum is 1000. The default is 100. For more information, see Paginating Results.}

\item{page_number}{(Optional) The offset for paging. The default is 1. For more information, see Paginating Results.}

\item{include_metadata}{Logical indicating whether to include metadata columns in the result (default: FALSE).}
}
\value{
A table containing the groups for the specified user.
}
\description{
Gets a list of groups that the specified user is a member of.
}
\seealso{
Other Tableau REST API: 
\code{\link{add_tags_to_view}()},
\code{\link{add_tags_to_workbook}()},
\code{\link{add_user_to_group}()},
\code{\link{add_user_to_site}()},
\code{\link{authenticate_server}()},
\code{\link{download_filtered_tableau_image}()},
\code{\link{download_tableau_crosstab_excel}()},
\code{\link{download_tableau_data}()},
\code{\link{download_workbooks_server_pdf}()},
\code{\link{download_workbooks_server_powerpoint}()},
\code{\link{download_workbooks_server}()},
\code{\link{get_server_connected_apps}()},
\code{\link{get_server_datasources}()},
\code{\link{get_server_groups}()},
\code{\link{get_server_info}()},
\code{\link{get_server_jobs}()},
\code{\link{get_server_projects}()},
\code{\link{get_server_refresh_tasks}()},
\code{\link{get_server_schedules}()},
\code{\link{get_server_user_favorites}()},
\code{\link{get_server_users}()},
\code{\link{get_server_views}()},
\code{\link{get_server_workbooks}()},
\code{\link{get_table_assets}()},
\code{\link{get_users_in_group}()},
\code{\link{remove_user_from_group}()},
\code{\link{remove_user_from_site}()}
}
\concept{Tableau REST API}
