% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hierarchy.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Get Workbook drilldown hierarchies}
\usage{
get_hierarchy(wb)
}
\arguments{
\item{wb}{The path to the Tableau workbook [.twb].}
}
\value{
Dataframe containing the workbook hierarchy drill downs.
}
\description{
Get Workbook drilldown hierarchies
}
\seealso{
Other xml: 
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
