% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodenames.R
\name{get_nodenames}
\alias{get_nodenames}
\title{Get all nodenames}
\usage{
get_nodenames(proc)
}
\arguments{
\item{proc}{The rootnode}
}
\value{
nodenames
}
\description{
Get all nodenames
}
\seealso{
Other xml: 
\code{\link{check_dashboard_names}()},
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
