% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameter}
\alias{get_parameter}
\title{Get all parameters in workbook.}
\usage{
get_parameter(wb)
}
\arguments{
\item{wb}{The path to the tableau workbook [.twb].}
}
\value{
Dataframe
}
\description{
Get all parameters in workbook.
}
\seealso{
Other xml: 
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
