% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server_location.R
\name{get_server_location}
\alias{get_server_location}
\title{Get the server location.}
\usage{
get_server_location(wb)
}
\arguments{
\item{wb}{The path to the Tableau workbook [.twb].}
}
\value{
The server location of the workbook.
}
\description{
Get the server location.
}
\seealso{
Other xml: 
\code{\link{adjust_tableau_size}()},
\code{\link{check_dashboard_names}()},
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
