% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tableau_data_source.R
\name{get_tableau_data_source}
\alias{get_tableau_data_source}
\title{Get Tableau data source}
\usage{
get_tableau_data_source(dashboard)
}
\arguments{
\item{dashboard}{Path to Tableau twb file.}
}
\value{
Data source of the workbook
}
\description{
Search XML for data source of Tableau workbook
}
\examples{
\dontrun{
# Get Tableau data source
data_source <- get_tableau_data_source(dashboard = "path/to/workbook.twb")
print(data_source)
}
}
\seealso{
Other xml: 
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
