% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_in_group.R
\name{get_users_in_group}
\alias{get_users_in_group}
\title{Get Users in Group}
\usage{
get_users_in_group(
  tableau,
  api_version = 3.19,
  group_id,
  page_size = 100,
  page_number = 1
)
}
\arguments{
\item{tableau}{A list containing the Tableau authentication variables: `base_url`, `token`, `user_id`, and `site_id`.}

\item{api_version}{The version of the API to use, such as 3.21.}

\item{group_id}{The ID of the group to get the users for.}

\item{page_size}{The number of items to return in one response. The minimum is 1. The maximum is 1000. The default is 100.}

\item{page_number}{The offset for paging. The default is 1.}
}
\value{
A list of users in the specified group.
}
\description{
Gets a list of users in the specified group.
}
\seealso{
Other Tableau REST API: 
\code{\link{add_tags_to_view}()},
\code{\link{add_tags_to_workbook}()},
\code{\link{add_user_to_group}()},
\code{\link{add_user_to_site}()},
\code{\link{authenticate_PAT}()},
\code{\link{authenticate_server}()},
\code{\link{create_group}()},
\code{\link{delete_group}()},
\code{\link{download_datasource}()},
\code{\link{download_filtered_tableau_image}()},
\code{\link{download_tableau_crosstab_excel}()},
\code{\link{download_tableau_data}()},
\code{\link{download_workbooks_server_pdf}()},
\code{\link{download_workbooks_server_powerpoint}()},
\code{\link{download_workbooks_server}()},
\code{\link{get_groups_for_user}()},
\code{\link{get_mobile_security_settings}()},
\code{\link{get_server_connected_apps}()},
\code{\link{get_server_datasources}()},
\code{\link{get_server_groups}()},
\code{\link{get_server_info}()},
\code{\link{get_server_jobs}()},
\code{\link{get_server_projects}()},
\code{\link{get_server_refresh_tasks}()},
\code{\link{get_server_schedules}()},
\code{\link{get_server_user_favorites}()},
\code{\link{get_server_users}()},
\code{\link{get_server_views}()},
\code{\link{get_server_workbooks}()},
\code{\link{get_table_assets}()},
\code{\link{query_user_on_site}()},
\code{\link{remove_user_from_group}()},
\code{\link{remove_user_from_site}()},
\code{\link{update_group}()},
\code{\link{update_user}()}
}
\concept{Tableau REST API}
