% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workbook_id.R
\name{get_workbook_id}
\alias{get_workbook_id}
\title{Get workbook ID name.}
\usage{
get_workbook_id(wb)
}
\arguments{
\item{wb}{The path to the Tableau workbook [.twb].}
}
\value{
The ID name of the workbook.
}
\description{
Get workbook ID name.
}
\seealso{
Other xml: 
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_tabs}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
