% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workbook_tabs.R
\name{get_workbook_tabs}
\alias{get_workbook_tabs}
\title{Get workbook tabs}
\usage{
get_workbook_tabs(wb)
}
\arguments{
\item{wb}{The path to the Tableau workbook file [.twb].}
}
\value{
Dataframe
}
\description{
Get workbook tabs
}
\seealso{
Other xml: 
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{make_rootnodes}()}
}
\concept{xml}
