% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workbook_urls.R
\name{get_workbook_urls}
\alias{get_workbook_urls}
\title{Get Workbook URLs}
\usage{
get_workbook_urls(wb)
}
\arguments{
\item{wb}{The path to the Tableau workbook file (.twb).}
}
\value{
A data frame containing the URLs found in the workbook.
}
\description{
This function parses an XML document representing a Tableau workbook and finds all nodes that have a url attribute containing 'views'.
It then extracts these URLs and returns them in a data frame.
}
