% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rootnodes.R
\name{make_rootnodes}
\alias{make_rootnodes}
\title{make_rootnodes}
\usage{
make_rootnodes(wb)
}
\arguments{
\item{wb}{The path to the tableau workbook file [.twb].}
}
\value{
The rootnodes.
}
\description{
make_rootnodes
}
\seealso{
Other xml: 
\code{\link{check_dashboard_names}()},
\code{\link{get_actions}()},
\code{\link{get_folders}()},
\code{\link{get_hierarchy}()},
\code{\link{get_nodenames}()},
\code{\link{get_parameter}()},
\code{\link{get_revision}()},
\code{\link{get_server_location}()},
\code{\link{get_tableau_data_source}()},
\code{\link{get_variable_folders}()},
\code{\link{get_variables}()},
\code{\link{get_workbook_id}()},
\code{\link{get_workbook_tabs}()}
}
\concept{xml}
