% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view_filtered.R
\name{row_to_name}
\alias{row_to_name}
\title{Concat row to name}
\usage{
row_to_name(row)
}
\arguments{
\item{row}{row of dataframe}
}
\value{
namr of the row
}
\description{
Concat row to name
}
