% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view_filtered.R
\name{row_to_query}
\alias{row_to_query}
\title{Dataframe row to query}
\usage{
row_to_query(row)
}
\arguments{
\item{row}{row of dataframe}
}
\value{
Query to the api call
}
\description{
Dataframe row to query
}
