% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_server.R
\name{tableau_base_url}
\alias{tableau_base_url}
\title{Get the Tableau base URL from the environment variable}
\usage{
tableau_base_url()
}
\value{
The Tableau base URL stored in the TABLEAU_BASE_URL environment variable.
}
\description{
Get the Tableau base URL from the environment variable
}
