% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_PAT.R
\name{tableau_content_url}
\alias{tableau_content_url}
\title{Get the Tableau content URL from the environment variable}
\usage{
tableau_content_url()
}
\value{
The Tableau content URL stored in the TABLEAU_CONTENT_URL environment variable.
}
\description{
Get the Tableau content URL from the environment variable
}
