% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_server.R
\name{tableau_password}
\alias{tableau_password}
\title{Get the Tableau password from the environment variable}
\usage{
tableau_password()
}
\value{
The Tableau password stored in the TABLEAU_PASSWORD environment variable.
}
\description{
Get the Tableau password from the environment variable
}
