% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_PAT.R
\name{tableau_pat_name}
\alias{tableau_pat_name}
\title{Get the Tableau PAT name from the environment variable}
\usage{
tableau_pat_name()
}
\value{
The Tableau PAT name stored in the TABLEAU_PAT_NAME environment variable.
}
\description{
Get the Tableau PAT name from the environment variable
}
