% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_PAT.R
\name{tableau_pat_secret}
\alias{tableau_pat_secret}
\title{Get the Tableau PAT secret from the environment variable}
\usage{
tableau_pat_secret()
}
\value{
The Tableau PAT secret stored in the TABLEAU_PAT_SECRET environment variable.
}
\description{
Get the Tableau PAT secret from the environment variable
}
