% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_server.R
\name{tableau_username}
\alias{tableau_username}
\title{Get the Tableau username from the environment variable}
\usage{
tableau_username()
}
\value{
The Tableau username stored in the TABLEAU_USERNAME environment variable.
}
\description{
Get the Tableau username from the environment variable
}
