% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view_filtered.R
\name{download_filtered_tableau_image}
\alias{download_filtered_tableau_image}
\title{Download filtered tableau views to png images from dataframe}
\usage{
download_filtered_tableau_image(
  df,
  base_url,
  site_id,
  token,
  view_id,
  path_to_save
)
}
\arguments{
\item{df}{Dataframe containing filter columns and filter values}

\item{base_url}{Tableau server url}

\item{site_id}{The site id}

\item{token}{API token}

\item{view_id}{The id of the view to download}

\item{path_to_save}{The directory to write the images to}
}
\value{
NULL.
}
\description{
Download filtered tableau views to png images from dataframe
}
