% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view_filtered.R
\name{download_filtered_tableau_image}
\alias{download_filtered_tableau_image}
\title{Download filtered Tableau views to PNG images from a dataframe.}
\usage{
download_filtered_tableau_image(
  tableau,
  df,
  view_id,
  path_to_save,
  api_version = "3.8"
)
}
\arguments{
\item{tableau}{A list containing the Tableau authentication variables: `base_url`, `token`, `user_id`, and `site_id`.}

\item{df}{Dataframe containing filter columns and filter values.}

\item{view_id}{The ID of the view to download.}

\item{path_to_save}{The directory to write the images to.}

\item{api_version}{The API version to use (default: 3.8).}
}
\description{
Downloads PNG images of filtered Tableau views based on the provided dataframe containing filter columns and filter values.
}
\seealso{
Other Tableau REST API: 
\code{\link{add_tags_to_view}()},
\code{\link{add_tags_to_workbook}()},
\code{\link{add_user_to_group}()},
\code{\link{add_user_to_site}()},
\code{\link{authenticate_server}()},
\code{\link{download_tableau_crosstab_excel}()},
\code{\link{download_tableau_data}()},
\code{\link{download_workbooks_server_pdf}()},
\code{\link{download_workbooks_server_powerpoint}()},
\code{\link{download_workbooks_server}()},
\code{\link{get_server_connected_apps}()},
\code{\link{get_server_datasources}()},
\code{\link{get_server_groups}()},
\code{\link{get_server_info}()},
\code{\link{get_server_jobs}()},
\code{\link{get_server_projects}()},
\code{\link{get_server_refresh_tasks}()},
\code{\link{get_server_schedules}()},
\code{\link{get_server_user_favorites}()},
\code{\link{get_server_users}()},
\code{\link{get_server_views}()},
\code{\link{get_server_workbooks}()},
\code{\link{get_table_assets}()},
\code{\link{get_users_in_group}()},
\code{\link{remove_user_from_group}()},
\code{\link{remove_user_from_site}()}
}
\concept{Tableau REST API}
