% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tableau_data.R
\name{download_tableau_data}
\alias{download_tableau_data}
\title{Download Tableau view data as Excel}
\usage{
download_tableau_data(base_url, site_id, token, view_id, path_to_save)
}
\arguments{
\item{base_url}{Tableau server URL}

\item{site_id}{The site ID}

\item{token}{API token}

\item{view_id}{The ID of the view to download}

\item{path_to_save}{The directory to write the data Excel file to}

\item{api_version}{The API version; default set to 3.8}
}
\description{
Download Tableau view data as Excel
}
